backend/python/requirements.txt
fastapi
uvicorn
chromadb
zai-sdk
python-dotenv
sqlalchemy
requests
sentence-transformers
python-docx
python-multipart
tiktoken
pdfplumber
